@@ -90,7 +90,20 @@ class ServerError extends FancyException {
   final ErrorBody body;
 }
 
+// 3xx
+
+class MovedPermanentlyError extends ServerError {
+  MovedPermanentlyError(ErrorBody body, this.location)
+      : super(body, (_) {
+          throw Exception('MovedPermanentlyError should never be shown to the '
+              'user.');
+        });
+
+  final String location;
+}
+
 // 4xx
+
 class BadRequestError extends ServerError {
   BadRequestError(ErrorBody body)
       : super(body, (context) => context.s.app_error_badRequest);
@@ -121,14 +134,15 @@ class ConflictError extends ServerError {
 }
 
 class TooManyRequestsError extends ServerError {
-  TooManyRequestsError(ErrorBody body, {@required this.timeToWait})
-      : assert(timeToWait != null),
-        super(body, (context) => context.s.app_error_rateLimit(timeToWait));
+  TooManyRequestsError(ErrorBody body, {@required this.durationToWait})
+      : assert(durationToWait != null),
+        super(body, (context) => context.s.app_error_rateLimit(durationToWait));
 
-  final Duration timeToWait;
+  final Duration durationToWait;
 }
 
 // 5xx
+
 class InternalServerError extends ServerError {
   InternalServerError(ErrorBody body)
       : super(body, (context) => context.s.app_error_internal);
@@ -223,8 +237,8 @@ class NetworkService {
     }
     services.banners.remove(Banners.offline);
 
-    // Succeed if its a 2xx or 3xx status code.
-    if (response.statusCode ~/ 100 == 2 || response.statusCode ~/ 100 == 3) {
+    // Succeed if its a 2xx status code.
+    if (response.statusCode ~/ 100 == 2) {
       services.banners.remove(Banners.tokenExpired);
       return response;
     }
@@ -272,16 +286,10 @@ class NetworkService {
     }
 
     if (response.statusCode == HttpStatus.tooManyRequests) {
-      final timeToWait = () {
-        try {
-          return Duration(
-            seconds: error.data['timeToWait'],
-          );
-        } catch (_) {
-          return Duration(seconds: 10);
-        }
-      }();
-      throw TooManyRequestsError(error, timeToWait: timeToWait);
+      final secondsToWait = error.data['timeToWait'];
+      final durationToWait =
+          Duration(seconds: secondsToWait is int ? secondsToWait : 10);
+      throw TooManyRequestsError(error, durationToWait: durationToWait);
     }
 
     if (response.statusCode == HttpStatus.internalServerError) {
@@ -290,6 +298,10 @@ class NetworkService {
 
     if (response.statusCode == HttpStatus.notFound) {
       throw NotFoundError(error);
+    }
+
+    if (response.statusCode == HttpStatus.movedPermanently) {
+      throw MovedPermanentlyError(error, response.headers['location']);
     }
 
     throw UnimplementedError(
