@@ -43,7 +43,7 @@ class SortFilter<T> {
   }
 
   /// Parses a query string generated by the web client, e.g.
-  /// https://schul-cloud.org/homework/#?dueDateFrom=2020-03-09&dueDateTo=2020-03-27&private=true&publicSubmissions=false&sort=updatedAt&sortorder=1&teamSubmissions=true
+  /// https://hpi-schul-cloud.de/homework/#?dueDateFrom=2020-03-09&dueDateTo=2020-03-27&private=true&publicSubmissions=false&sort=updatedAt&sortorder=1&teamSubmissions=true
   SortFilterSelection<T> tryParseWebQuery(Map<String, String> query) {
     if (query.isEmpty) {
       return defaultSelection;
