@@ -3,7 +3,8 @@ import 'package:schulcloud/main.dart' as core;
 import 'main_sc.dart';
 
 final scTestAppConfig = scAppConfig.copyWith(
-  host: 'test.schul-cloud.org',
+  name: 'sc_test',
+  host: 'test.hpi-schul-cloud.org',
   title: 'HPI Schul-Cloud (Test)',
 );
 
